+++ packages/pipo-scripts/build.js
@@ -30,32 +30,32 @@ function getBabelConfig() {
   return require(config);
 }
 
-class OutputWebpackBuild {
-  // constructor(options) {
-  //   this.options = options;
-  // }
+// class OutputWebpackBuild {
+//   // constructor(options) {
+//   //   this.options = options;
+//   // }
 
-  apply(compiler) {
-    const { emit } = compiler.hooks;
-    // use `emit.tapPromise`, if I use an async function
-    emit.tap(this.constructor.name, (compilation) => {
-      compilation.modules.forEach((module) => {
-        module.dependencies.forEach((dependency) => {
-          if (dependency.type === 'harmony init') {
-            const filename = dependency.originModule.userRequest.replace(
-              process.cwd(),
-              ''
-            );
-            outputFileSync(
-              join('dist-webpack', `${filename}.js`),
-              dependency.originModule._source._value
-            );
-          }
-        });
-      });
-    });
-  }
-}
+//   apply(compiler) {
+//     const { emit } = compiler.hooks;
+//     // use `emit.tapPromise`, if I use an async function
+//     emit.tap(this.constructor.name, (compilation) => {
+//       compilation.modules.forEach((module) => {
+//         module.dependencies.forEach((dependency) => {
+//           if (dependency.type === 'harmony init') {
+//             const filename = dependency.originModule.userRequest.replace(
+//               process.cwd(),
+//               ''
+//             );
+//             outputFileSync(
+//               join('dist-webpack', `${filename}.js`),
+//               dependency.originModule._source._value
+//             );
+//           }
+//         });
+//       });
+//     });
+//   }
+// }
 
 const isWebApp = existsSync('src/index.html');
 
@@ -79,7 +79,13 @@ const config = {
       {
         test: /\.(js|jsx|ts|tsx)$/,
         exclude: /node_modules/,
-        use: [{ loader: 'babel-loader', options: getBabelConfig() }]
+        // include: join(cwd, 'src'),
+        use: [
+          {
+            loader: 'babel-loader',
+            options: getBabelConfig()
+          }
+        ]
       },
       {
         test: /\.css$/,
@@ -91,6 +97,22 @@ const config = {
           },
           { loader: 'css-loader' }
         ]
+      },
+      {
+        test: /\.(png|jpg|gif|svg)$/,
+        loader: 'url-loader?limit=1000&name=[name]-[hash].[ext]'
+      },
+      {
+        test: /\.eot(\?v=\d+\.\d+\.\d+)?$/,
+        loader: 'file-loader'
+      },
+      {
+        test: /\.(woff|woff2)$/,
+        loader: 'url-loader?prefix=font/&limit=5000'
+      },
+      {
+        test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/,
+        loader: 'url-loader?limit=10000&mimetype=application/octet-stream'
       }
     ]
   },
@@ -103,7 +125,15 @@ const config = {
   mode: 'production',
   // target: 'node', // trigger based on targets of babelrc?
   resolve: {
-    extensions: ['.ts', '.tsx', '.js', '.jsx']
+    extensions: ['.ts', '.tsx', '.js', '.jsx'],
+    mainFields: [
+      'webpack',
+      // defaults
+      // see https://github.com/webpack/webpack/blob/dc50c0360e87204ea77172910e877f8c510f3bfb/lib/WebpackOptionsDefaulter.js#L84
+      'browser',
+      'module',
+      'main'
+    ]
   },
   resolveLoader: {
     modules: [
@@ -114,7 +144,14 @@ const config = {
       // point to the same directory
       join(__dirname, '../../node_modules'),
       'node_modules'
-    ]
+    ],
+    alias: {
+      // see https://www.npmjs.com/package/copy-loader
+      ['copy-loader']: `file-loader?name=[path][name].[ext]&context=./${join(
+        cwd,
+        'src'
+      )}`
+    }
   },
   stats
 };